Moved click window packet and made ItemUtils
